requirements.txt
google-generativeai
gradio
feedparser
pandas
matplotlib
